Feat. change DB fields state into atom
